requirements.txt
obs-websocket-py==1.0
websocket-client==1.8.0
websockets==12.0
